move codewhispererFeatureConfigService to amazonq/shared
